Hungarian pipeline: QwenVL OCR, mask fixes, dedup, bubble merge improvements
Key changes:
- Add QwenVL (Qwen2.5-VL-3B) as primary OCR model with multi-OCR fallback scoring
- Fix inpainting blend: revert feathered Gaussian to hard binary mask
- Fix _apply_textlines_to_mask: Otsu minority-class detection, outline detection,
  background-only std for adaptive dilation kernel (5 vs 11px)
- OCR garbage filter: regex for VLM hallucination phrases
- TextBlock text dedup: substring elimination + suffix-prefix overlap trimming
  to prevent duplicated OCR fragments in merged regions
- Bubble force-merge: gap-based splitting (2.5x font_size threshold) to keep
  UI/game window title separate from description text
- textline_merge gamma: 0.5 -> 0.3 to reduce over-merging of nearby bubbles
- Verbose-only debug saves: mask_final, inpaint_input, inpainted, ocr_sources
  now only written when --verbose flag is set
- Per-region inpainting for tall pages (h > w*3) with correct paste-back
- Add HealTheWorld and MrHorror fonts
- Add ocr_vl_test config and QwenVL OCR model

+++ manga_translator/inpainting/inpainting_lama_mpe.py
@@ -114,7 +114,8 @@ class LamaMPEInpainter(OfflineInpainter):
             img_inpainted = ((img_inpainted_torch.cpu().squeeze_(0).permute(1, 2, 0).numpy() + 1.0) * 127.5).astype(np.uint8)
         if new_h != height or new_w != width:
             img_inpainted = cv2.resize(img_inpainted, (width, height), interpolation = cv2.INTER_LINEAR)
-        ans = img_inpainted * mask_original + img_original * (1 - mask_original)
+        ans = (img_inpainted.astype(np.float32) * mask_original +
+               img_original.astype(np.float32) * (1.0 - mask_original)).astype(np.uint8)
         return ans
